Lesson6

+++ app/controllers/questions_controller.rb
@@ -1,4 +1,6 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: [:index, :show]
+
   def index
     @questions = Question.all
   end
@@ -7,7 +9,7 @@ class QuestionsController < ApplicationController
   end
 
   def new
-#    @question = Question.new
+    @question = current_user.questions.new
   end
 
   def edit
@@ -15,8 +17,9 @@ class QuestionsController < ApplicationController
 
   def create
     if question.save
-      redirect_to question
+      redirect_to question, notice: 'Successfully added.'
     else
+      flash.now[:alert] = "Unable to add!"
       render :new
     end
   end
@@ -30,17 +33,27 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
+    if current_user&.author?(question)
+      question.destroy
+      flash[:notice] = 'Deleted successfully'
+    else
+      flash[:alert] = "You can not delete"
+    end
     redirect_to questions_path
   end
+
   private
 
   def question
-    @question ||= params[:id] ? Question.find(params[:id]) : Question.new(question_params)
+    @question ||= params[:id] ? Question.find(params[:id]) : current_user.questions.new(question_params)
   end
-
   helper_method :question
 
+  def answer
+    @answer ||= question.answers.new(user: current_user)
+  end
+  helper_method :answer
+  
   def question_params
     params.require(:question).permit(:title, :body)
   end
